Use the order number as BTCPay order id
Follow-up to #53.

Because the BTCPay order id shows up as a reference in the invoices list, its inconvenient for merchants to see the actual ID instead of the order number, which they expect.

These changes reset the part from the first PR, where the order ID got used consistently. Now we are setting the order number as the ID again, to suit the merchants need for easier bookkeeping.

To accommodate for that, now the `ipn_callback` gets the actual order ID from the PoS data that's been set on invoice creation by the `process_payment` method.

This way we can hide the complexity without the need for a custom filter function to convert the order number to order id. This has been error prone and triggered the first PR.